Implement piano-style mapper
ref #1
